requirements.txt
aiofiles==23.2.1
aiogram==3.11.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.7.4
frozenlist==1.4.1
idna==3.7
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
tortoise-orm==0.21.5
typing_extensions==4.12.2
yarl==1.9.4
